Fixed non-whitelisted ionicons log message
Fixes #4514

The _ionicons-font.less file was referencing the font files in a relative
way, but the font files weren't in the directory with it. The reason it
worked at all was because there's a separate stylesheet for fonts in the
font directory that gave the real location of the ionicons font files.